Refactor ImageBlenders to share a common base class.
Extracted duplicated bounding-box calculation, bounds safety, and array-span setup from the four individual `IImageBlender` implementations (`DirectImageBlender`, `AlphaImageBlender`, `AlphaOnlyImageBlender`, `RGBOnlyImageBlender`) into a shared `ImageBlenderBase`. Subclasses now override the `BlendInternal` method to provide the specific inner-loop blending logic. This significantly reduces code duplication and applies an optimization (`toPos += 4` rather than array lookup logic over every single pixel) consistently across all blenders.
