+++ Eede.Domain/ImageEditing/Blending/DirectImageBlender.cs
@@ -1,40 +1,15 @@
 #nullable enable
-using Eede.Domain.SharedKernel;
 using System;
 
 namespace Eede.Domain.ImageEditing.Blending;
 
-public class DirectImageBlender : IImageBlender
+public class DirectImageBlender : ImageBlenderBase
 {
-    public Picture Blend(Picture from, Picture to)
+    protected override void BlendInternal(ReadOnlySpan<byte> fromSpan, byte[] toPixels, int fromPos, int toPos)
     {
-        return Blend(from, to, new Position(0, 0));
-    }
-
-    public Picture Blend(Picture from, Picture to, Position toPosition)
-    {
-        byte[] toPixels = to.CloneImage();
-
-        int startY = Math.Max(0, toPosition.Y);
-        int startX = Math.Max(0, toPosition.X);
-        int maxY = Math.Min(toPosition.Y + from.Height, to.Height);
-        int maxX = Math.Min(toPosition.X + from.Width, to.Width);
-
-        ReadOnlySpan<byte> fromSpan = from.AsSpan();
-        for (int y = startY; y < maxY; y++)
-        {
-            int toPos = (startX * 4) + (to.Stride * y);
-            int fromPos = ((startX - toPosition.X) * 4) + (from.Stride * (y - toPosition.Y));
-            for (int x = startX; x < maxX; x++)
-            {
-                toPixels[toPos + 0] = fromSpan[fromPos + 0];
-                toPixels[toPos + 1] = fromSpan[fromPos + 1];
-                toPixels[toPos + 2] = fromSpan[fromPos + 2];
-                toPixels[toPos + 3] = fromSpan[fromPos + 3];
-                toPos += 4;
-                fromPos += 4;
-            }
-        }
-        return Picture.Create(to.Size, toPixels);
+        toPixels[toPos + 0] = fromSpan[fromPos + 0];
+        toPixels[toPos + 1] = fromSpan[fromPos + 1];
+        toPixels[toPos + 2] = fromSpan[fromPos + 2];
+        toPixels[toPos + 3] = fromSpan[fromPos + 3];
     }
 }
